Moved shader load to ClientProxy with SideOnly, and changed loading to getResourceAsStream with rooted absolute path.
